Refactor everything out into chat_template_utils
