+++ src/transformers/utils/chat_template_utils.py
@@ -15,7 +15,17 @@
 import inspect
 import json
 import re
-from typing import Any, Callable, Dict, Optional, Tuple, Union, get_args, get_origin, get_type_hints
+from contextlib import contextmanager
+from datetime import datetime
+from functools import lru_cache
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union, get_args, get_origin, get_type_hints
+
+import jinja2
+from jinja2 import nodes
+from jinja2.exceptions import TemplateError
+from jinja2.ext import Extension
+from jinja2.sandbox import ImmutableSandboxedEnvironment
+from packaging import version
 
 
 BASIC_TYPES = (int, float, str, bool, Any, type(None), ...)
@@ -314,3 +324,91 @@ def get_json_schema(func: Callable) -> Dict:
     if return_dict is not None:
         output["return"] = return_dict
     return {"type": "function", "function": output}
+
+
+class AssistantTracker(Extension):
+    # This extension is used to track the indices of assistant-generated tokens in the rendered chat
+    tags = {"generation"}
+
+    def __init__(self, environment: ImmutableSandboxedEnvironment):
+        # The class is only initiated by jinja.
+        super().__init__(environment)
+        environment.extend(activate_tracker=self.activate_tracker)
+        self._rendered_blocks = None
+        self._generation_indices = None
+
+    def parse(self, parser: jinja2.parser.Parser) -> jinja2.nodes.CallBlock:
+        lineno = next(parser.stream).lineno
+        body = parser.parse_statements(["name:endgeneration"], drop_needle=True)
+        return nodes.CallBlock(self.call_method("_generation_support"), [], [], body).set_lineno(lineno)
+
+    @jinja2.pass_eval_context
+    def _generation_support(self, context: jinja2.nodes.EvalContext, caller: jinja2.runtime.Macro) -> str:
+        rv = caller()
+        if self.is_active():
+            # Only track generation indices if the tracker is active
+            start_index = len("".join(self._rendered_blocks))
+            end_index = start_index + len(rv)
+            self._generation_indices.append((start_index, end_index))
+        return rv
+
+    def is_active(self) -> bool:
+        return self._rendered_blocks or self._generation_indices
+
+    @contextmanager
+    def activate_tracker(self, rendered_blocks: List[int], generation_indices: List[int]):
+        try:
+            if self.is_active():
+                raise ValueError("AssistantTracker should not be reused before closed")
+            self._rendered_blocks = rendered_blocks
+            self._generation_indices = generation_indices
+
+            yield
+        finally:
+            self._rendered_blocks = None
+            self._generation_indices = None
+
+
+def _render_with_assistant_indices(
+    compiled_template, messages, tools, documents, add_generation_prompt, **template_kwargs
+):
+    rendered_blocks = []
+    generation_indices = []
+    with compiled_template.environment.activate_tracker(rendered_blocks, generation_indices):
+        for block in compiled_template.generate(
+            messages=messages,
+            tools=tools,
+            documents=documents,
+            add_generation_prompt=add_generation_prompt,
+            **template_kwargs,
+        ):
+            rendered_blocks.append(block)
+        rendered_chat = "".join(rendered_blocks)
+    return rendered_chat, generation_indices
+
+
+@lru_cache
+def _compile_jinja_template(chat_template):
+    if version.parse(jinja2.__version__) < version.parse("3.1.0"):
+        raise ImportError(
+            "apply_chat_template requires jinja2>=3.1.0 to be installed. Your version is " f"{jinja2.__version__}."
+        )
+
+    def raise_exception(message):
+        raise TemplateError(message)
+
+    def tojson(x, ensure_ascii=False, indent=None, separators=None, sort_keys=False):
+        # We override the built-in tojson filter because Jinja's default filter escapes HTML characters
+        # We also expose some options like custom indents and separators
+        return json.dumps(x, ensure_ascii=ensure_ascii, indent=indent, separators=separators, sort_keys=sort_keys)
+
+    def strftime_now(format):
+        return datetime.now().strftime(format)
+
+    jinja_env = ImmutableSandboxedEnvironment(
+        trim_blocks=True, lstrip_blocks=True, extensions=[AssistantTracker, jinja2.ext.loopcontrols]
+    )
+    jinja_env.filters["tojson"] = tojson
+    jinja_env.globals["raise_exception"] = raise_exception
+    jinja_env.globals["strftime_now"] = strftime_now
+    return jinja_env.from_string(chat_template)
